Implement user-defined keys for columns
Added API for setting a key identifier for a column. This value can be used to retrieve the column later with Grid::getColumnByKey.
Public API for using column-ids is removed, and only keys should be used by the users. Ids are only for internal implementation.
Also, parentColumns is renamed to topLevelColumns to avoid confusion. It's now lazily initialized when/if needed instead of keeping it always up to date, for memory/performance. Public getter of that is also removed, because it doesn't seem very useful.
Also, improved javadocs of Grid::getColumns to mention that the order may be incorrect when user-reordering is enabled.