Implement user-defined keys for columns (#27)
Added API for setting a key identifier for a column. This value can be used to retrieve the column later with Grid::getColumnByKey.

Public API for using column-ids is removed, and only keys should be used by the users. Ids are only for internal implementation.

Also, parentColumns are now accessible through the getTopLevelColumns method and its visibility has been changed to private.

Also, improved javadocs of Grid::getColumns to mention that the order may be incorrect when user-reordering is enabled.