Allow disabling item details opening by clicking on rows
Add getter and setter for detailsVisibleOnClick, which has a default value of true. The demo for details row is separated to it's own tab and there are now two demos: the first one demonstrates only the default functionality with details opening by clicking on rows, and the second one disables this and opens details with the server-side API. Also, changed mentions of "details row" to "item details" in the demo, since this is the API naming we're using.