Fix multiselection bug
There was a bug with grid in multiselection mode: The first click on a checkbox in the selection column would do the selection but the checkbox would stay unchecked. The second click would make the checkbox checked, but wouldn't change the selection. The same with unselecting.
This patch fixes the inconsistent state between checkboxes and selected items.