Add API for getting ContextMenu target item
This is implemented by adding a method to Grid which throws an exception
if there is no opened ContextMenu present. The API is not optimal, as it
would be nice to get the target item from the ContextMenu's click
events, but this is done to avoid having dependencies between Grid and
ContextMenu.