Fix setting initial sort order
Updating the client-side sorters needed to be postponed, because the
sorter elements were not present at the time of this JS execution when
done on the same roundtrip as creating/attaching the grid. The grid
would finalize its initialization afterwards with empty sort orders,
causing un-sorted data to be requested from the server.

Also, the initial call to the sorter observer needed to be disabled so
it doesn't reset the sort order.

Fix #13