DetailsRenderer and ClassNameGenerator in the same Grid fails
If I have a column with a ClassNameGenerator in a Grid where I also have a DetailsRenderer,
The code accesses the ClassNameGenerator-method with an undefined column. This is a solution to that problem.
It might be a bandaid on another problem, but it at least solves it.
(This is my first pull-request ever, tell me if I do something wrong)