Remove an invalid test
The removed test is asserting that after removing a row, it no longer contains row details content. The test conflicts with the PR's fix that causes component renderers of removed rows to no longer to update. If a row is removed, physically the `<tr>` element gets `hidden`, and its content should no longer matter because none of it is rendered anyway.