Send null to client instead of {key:"null"} object (#815)
Grid#callSelectionFunctionForItems prepares items as a json array before sending it to connector. It creates an array of {key: KEY} object. When Grid#deselectAll() is called, the former method creates an object {key:"null"} and that object ends up being the activeItem on client. What this changes does it to send null to be set on client instead.