Replace Lambda with inline SerializableComparator
Since setComparator() uses Comparator as type parameter, Lambda will not be serializable. Inline SerlializableComparator class should be used instead. It would be even better to fix setComparator type parameter to be SerializableCompator, but there are additional problems to be solved before it can be made, as seen in https://github.com/vaadin/vaadin-grid-flow/pull/840

Fixes: https://github.com/vaadin/vaadin-grid-flow/issues/839