Add API to control nested properties null behavior
If nested properties have null in between, NPE is thrown. Depending on the application this may be undesirable. In some applications null's are not supposed to be there and that should be endorsed by throwing NPE, so the current behavior is correct for those. However in some cases it is desired to be more relaxed and properties with nested null's should be silently translated to null value instead.

Fixes: https://github.com/vaadin/flow/issues/8891