fix: prevent NPE for client-side only grids
fix #1077

`ItemCache#ensureSubCacheForScaledIndex` method is changed by
gridConnector, so it will make any grid on the page use the new
implementation. That leads to a call from undefined object in case of a
grid entirely defined on the client side needs to call this method.

This change stores the original implementation and calls it when the
grid doesn't have a `$connector` object.