fix: make connector patch ItemCache only once
Previous fix for client and server grids on the same page was to
store the original `ensureSubCacheForScaledIndex` method before it
was patched by connector, but that introduced a stack overflow
error after a second grid is added from the server.

This fix makes ItemCache patching to happen only once the first
server grid is added, and prevents it from happening after other
grids are added.

Fixes #1077